Scripts/Commands: Convert character and server commands to commandscript
Closes #6856
