+++ src/server/game/Chat/Chat.h
@@ -119,10 +119,13 @@ class ChatHandler
 
         GameObject* GetNearbyGameObject();
         GameObject* GetObjectGlobalyWithGuidOrNearWithDbGuid(uint32 lowguid, uint32 entry);
-        bool HasSentErrorMessage() const { return sentErrorMessage;}
-        void SetSentErrorMessage(bool val){ sentErrorMessage = val;};
-        static bool LoadCommandTable() { return load_command_table;}
-        static void SetLoadCommandTable(bool val){ load_command_table = val;};
+        bool HasSentErrorMessage() const { return sentErrorMessage; }
+        void SetSentErrorMessage(bool val){ sentErrorMessage = val; }
+        static bool LoadCommandTable() { return load_command_table; }
+        static void SetLoadCommandTable(bool val) { load_command_table = val; }
+        
+        // cs_character
+        void HandleCharacterLevel(Player* player, uint64 playerGuid, uint32 oldLevel, uint32 newLevel);
 
     protected:
         explicit ChatHandler() : m_session(NULL) {}      // for CLI subclass
@@ -146,19 +149,6 @@ class ChatHandler
         bool HandleBanListAccountCommand(const char* args);
         bool HandleBanListCharacterCommand(const char* args);
         bool HandleBanListIPCommand(const char* args);
-
-        bool HandleCharacterCustomizeCommand(const char* args);
-        bool HandleCharacterChangeFactionCommand(const char* args);
-        bool HandleCharacterChangeRaceCommand(const char * args);
-        bool HandleCharacterDeletedDeleteCommand(const char* args);
-        bool HandleCharacterDeletedListCommand(const char* args);
-        bool HandleCharacterDeletedRestoreCommand(const char* args);
-        bool HandleCharacterDeletedOldCommand(const char* args);
-        bool HandleCharacterEraseCommand(const char* args);
-        bool HandleCharacterLevelCommand(const char* args);
-        bool HandleCharacterRenameCommand(const char* args);
-        bool HandleCharacterReputationCommand(const char* args);
-        bool HandleCharacterTitlesCommand(const char* args);
 
         bool HandleChannelSetOwnership(const char *args);
 
@@ -211,24 +201,6 @@ class ChatHandler
         bool HandleSendMailCommand(const char* args);
         bool HandleSendMessageCommand(const char * args);
         bool HandleSendMoneyCommand(const char* args);
-
-        bool HandleServerCorpsesCommand(const char* args);
-        bool HandleServerExitCommand(const char* args);
-        bool HandleServerIdleRestartCommand(const char* args);
-        bool HandleServerIdleShutDownCommand(const char* args);
-        bool HandleServerInfoCommand(const char* args);
-        bool HandleServerMotdCommand(const char* args);
-        bool HandleServerPLimitCommand(const char* args);
-        bool HandleServerRestartCommand(const char* args);
-        bool HandleServerSetLogLevelCommand(const char* args);
-        bool HandleServerSetMotdCommand(const char* args);
-        bool HandleServerShutDownCommand(const char* args);
-        bool HandleServerShutDownCancelCommand(const char* args);
-        bool HandleServerSetClosedCommand(const char* args);
-        bool HandleServerToggleQueryLogging(const char* args);
-
-        bool HandleServerSetLogFileLevelCommand(const char* args);
-        bool HandleServerSetDiffTimeCommand(const char* args);
 
         bool HandleUnBanAccountCommand(const char* args);
         bool HandleUnBanAccountByCharCommand(const char* args);
@@ -342,24 +314,7 @@ class ChatHandler
         bool HandleBanHelper(BanMode mode, char const* args);
         bool HandleBanInfoHelper(uint32 accountid, char const* accountname);
         bool HandleUnBanHelper(BanMode mode, char const* args);
-        void HandleCharacterLevel(Player* player, uint64 playerGuid, uint32 oldLevel, uint32 newLevel);
         void HandleLearnSkillRecipesHelper(Player* player, uint32 skill_id);
-
-        // Stores informations about a deleted character
-        struct DeletedInfo
-        {
-            uint32      lowguid;                            ///< the low GUID from the character
-            std::string name;                               ///< the character name
-            uint32      accountId;                          ///< the account id
-            std::string accountName;                        ///< the account name
-            time_t      deleteDate;                         ///< the date at which the character has been deleted
-        };
-
-        typedef std::list<DeletedInfo> DeletedInfoList;
-        bool GetDeletedCharacterInfoList(DeletedInfoList& foundList, std::string searchString = "");
-        std::string GenerateDeletedCharacterGUIDsWhereStr(DeletedInfoList::const_iterator& itr, DeletedInfoList::const_iterator const& itr_end);
-        void HandleCharacterDeletedListHelper(DeletedInfoList const& foundList);
-        void HandleCharacterDeletedRestoreHelper(DeletedInfo const& delInfo);
 
     private:
         bool _HandleGMTicketResponseAppendCommand(const char* args, bool newLine);
