+++ src/server/game/Chat/Commands/Level2.cpp
@@ -456,190 +456,6 @@ bool ChatHandler::HandlePInfoCommand(const char* args)
     return true;
 }
 
-//rename characters
-bool ChatHandler::HandleCharacterRenameCommand(const char* args)
-{
-    Player* target;
-    uint64 targetGuid;
-    std::string targetName;
-    if (!extractPlayerTarget((char*)args, &target, &targetGuid, &targetName))
-        return false;
-
-    if (target)
-    {
-        // check online security
-        if (HasLowerSecurity(target, 0))
-            return false;
-
-        PSendSysMessage(LANG_RENAME_PLAYER, GetNameLink(target).c_str());
-        target->SetAtLoginFlag(AT_LOGIN_RENAME);
-    }
-    else
-    {
-        // check offline security
-        if (HasLowerSecurity(NULL, targetGuid))
-            return false;
-
-        std::string oldNameLink = playerLink(targetName);
-
-        PSendSysMessage(LANG_RENAME_PLAYER_GUID, oldNameLink.c_str(), GUID_LOPART(targetGuid));
-
-        PreparedStatement* stmt = CharacterDatabase.GetPreparedStatement(CHAR_UPD_ADD_AT_LOGIN_FLAG);
-
-        stmt->setUInt16(0, uint16(AT_LOGIN_RENAME));
-        stmt->setUInt32(1, GUID_LOPART(targetGuid));
-
-        CharacterDatabase.Execute(stmt);
-    }
-
-    return true;
-}
-
-// customize characters
-bool ChatHandler::HandleCharacterCustomizeCommand(const char* args)
-{
-    Player* target;
-    uint64 targetGuid;
-    std::string targetName;
-    if (!extractPlayerTarget((char*)args, &target, &targetGuid, &targetName))
-        return false;
-
-    PreparedStatement* stmt = CharacterDatabase.GetPreparedStatement(CHAR_UPD_ADD_AT_LOGIN_FLAG);
-
-    stmt->setUInt16(0, uint16(AT_LOGIN_CUSTOMIZE));
-
-    if (target)
-    {
-        PSendSysMessage(LANG_CUSTOMIZE_PLAYER, GetNameLink(target).c_str());
-        target->SetAtLoginFlag(AT_LOGIN_CUSTOMIZE);
-
-        stmt->setUInt32(1, target->GetGUIDLow());
-    }
-    else
-    {
-        std::string oldNameLink = playerLink(targetName);
-
-        stmt->setUInt32(1, GUID_LOPART(targetGuid));
-
-        PSendSysMessage(LANG_CUSTOMIZE_PLAYER_GUID, oldNameLink.c_str(), GUID_LOPART(targetGuid));
-    }
-
-    CharacterDatabase.Execute(stmt);
-
-    return true;
-}
-
-bool ChatHandler::HandleCharacterChangeFactionCommand(const char* args)
-{
-    Player* target;
-    uint64 targetGuid;
-    std::string targetName;
-
-    if (!extractPlayerTarget((char*)args, &target, &targetGuid, &targetName))
-        return false;
-
-    PreparedStatement* stmt = CharacterDatabase.GetPreparedStatement(CHAR_UPD_ADD_AT_LOGIN_FLAG);
-
-    stmt->setUInt16(0, uint16(AT_LOGIN_CHANGE_FACTION));
-
-    if (target)
-    {
-        PSendSysMessage(LANG_CUSTOMIZE_PLAYER, GetNameLink(target).c_str());
-        target->SetAtLoginFlag(AT_LOGIN_CHANGE_FACTION);
-
-        stmt->setUInt32(1, target->GetGUIDLow());
-    }
-    else
-    {
-        std::string oldNameLink = playerLink(targetName);
-
-        PSendSysMessage(LANG_CUSTOMIZE_PLAYER_GUID, oldNameLink.c_str(), GUID_LOPART(targetGuid));
-
-        stmt->setUInt32(1, GUID_LOPART(targetGuid));
-    }
-
-    CharacterDatabase.Execute(stmt);
-
-    return true;
-}
-
-bool ChatHandler::HandleCharacterChangeRaceCommand(const char * args)
-{
-    Player* target;
-    uint64 targetGuid;
-    std::string targetName;
-    if (!extractPlayerTarget((char*)args, &target, &targetGuid, &targetName))
-        return false;
-
-    PreparedStatement* stmt = CharacterDatabase.GetPreparedStatement(CHAR_UPD_ADD_AT_LOGIN_FLAG);
-
-    stmt->setUInt16(0, uint16(AT_LOGIN_CHANGE_RACE));
-
-    if (target)
-    {
-        // TODO : add text into database
-        PSendSysMessage(LANG_CUSTOMIZE_PLAYER, GetNameLink(target).c_str());
-        target->SetAtLoginFlag(AT_LOGIN_CHANGE_RACE);
-
-        stmt->setUInt32(1, target->GetGUIDLow());
-    }
-    else
-    {
-        std::string oldNameLink = playerLink(targetName);
-
-        // TODO : add text into database
-        PSendSysMessage(LANG_CUSTOMIZE_PLAYER_GUID, oldNameLink.c_str(), GUID_LOPART(targetGuid));
-
-        stmt->setUInt32(1, GUID_LOPART(targetGuid));
-    }
-
-    CharacterDatabase.Execute(stmt);
-
-    return true;
-}
-
-bool ChatHandler::HandleCharacterReputationCommand(const char* args)
-{
-    Player* target;
-    if (!extractPlayerTarget((char*)args, &target))
-        return false;
-
-    LocaleConstant loc = GetSessionDbcLocale();
-
-    FactionStateList const& targetFSL = target->GetReputationMgr().GetStateList();
-    for (FactionStateList::const_iterator itr = targetFSL.begin(); itr != targetFSL.end(); ++itr)
-    {
-        const FactionState& faction = itr->second;
-        FactionEntry const* factionEntry = sFactionStore.LookupEntry(faction.ID);
-        char const* factionName = factionEntry ? factionEntry->name[loc] : "#Not found#";
-        ReputationRank rank = target->GetReputationMgr().GetRank(factionEntry);
-        std::string rankName = GetTrinityString(ReputationRankStrIndex[rank]);
-        std::ostringstream ss;
-        if (m_session)
-            ss << faction.ID << " - |cffffffff|Hfaction:" << faction.ID << "|h[" << factionName << ' ' << localeNames[loc] << "]|h|r";
-        else
-            ss << faction.ID << " - " << factionName << ' ' << localeNames[loc];
-
-        ss << ' ' << rankName << " (" << target->GetReputationMgr().GetReputation(factionEntry) << ')';
-
-        if (faction.Flags & FACTION_FLAG_VISIBLE)
-            ss << GetTrinityString(LANG_FACTION_VISIBLE);
-        if (faction.Flags & FACTION_FLAG_AT_WAR)
-            ss << GetTrinityString(LANG_FACTION_ATWAR);
-        if (faction.Flags & FACTION_FLAG_PEACE_FORCED)
-            ss << GetTrinityString(LANG_FACTION_PEACE_FORCED);
-        if (faction.Flags & FACTION_FLAG_HIDDEN)
-            ss << GetTrinityString(LANG_FACTION_HIDDEN);
-        if (faction.Flags & FACTION_FLAG_INVISIBLE_FORCED)
-            ss << GetTrinityString(LANG_FACTION_INVISIBLE_FORCED);
-        if (faction.Flags & FACTION_FLAG_INACTIVE)
-            ss << GetTrinityString(LANG_FACTION_INACTIVE);
-
-        SendSysMessage(ss.str().c_str());
-    }
-    return true;
-}
-
 bool ChatHandler::HandleLookupEventCommand(const char* args)
 {
     if (!*args)
@@ -843,13 +659,6 @@ bool ChatHandler::LookupPlayerSearchCommand(PreparedQueryResult result, int32 li
         return false;
     }
 
-    return true;
-}
-
-/// Triggering corpses expire check in world
-bool ChatHandler::HandleServerCorpsesCommand(const char* /*args*/)
-{
-    sObjectAccessor->RemoveOldCorpses();
     return true;
 }
 
@@ -1126,45 +935,5 @@ bool ChatHandler::HandleLookupTitleCommand(const char* args)
     }
     if (counter == 0)                                       // if counter == 0 then we found nth
         SendSysMessage(LANG_COMMAND_NOTITLEFOUND);
-    return true;
-}
-
-bool ChatHandler::HandleCharacterTitlesCommand(const char* args)
-{
-    if (!*args)
-        return false;
-
-    Player* target;
-    if (!extractPlayerTarget((char*)args, &target))
-        return false;
-
-    LocaleConstant loc = GetSessionDbcLocale();
-    char const* targetName = target->GetName();
-    char const* knownStr = GetTrinityString(LANG_KNOWN);
-
-    // Search in CharTitles.dbc
-    for (uint32 id = 0; id < sCharTitlesStore.GetNumRows(); id++)
-    {
-        CharTitlesEntry const* titleInfo = sCharTitlesStore.LookupEntry(id);
-        if (titleInfo && target->HasTitle(titleInfo))
-        {
-            std::string name = titleInfo->name[loc];
-            if (name.empty())
-                continue;
-
-            char const* activeStr = target && target->GetUInt32Value(PLAYER_CHOSEN_TITLE) == titleInfo->bit_index
-                ? GetTrinityString(LANG_ACTIVE)
-                : "";
-
-            char titleNameStr[80];
-            snprintf(titleNameStr, 80, name.c_str(), targetName);
-
-            // send title in "id (idx:idx) - [namedlink locale]" format
-            if (m_session)
-                PSendSysMessage(LANG_TITLE_LIST_CHAT, id, titleInfo->bit_index, id, titleNameStr, localeNames[loc], knownStr, activeStr);
-            else
-                PSendSysMessage(LANG_TITLE_LIST_CONSOLE, id, titleInfo->bit_index, name.c_str(), localeNames[loc], knownStr, activeStr);
-        }
-    }
     return true;
 }
