+++ src/server/game/Chat/Commands/Level3.cpp
@@ -2198,47 +2198,6 @@ void ChatHandler::HandleCharacterLevel(Player* player, uint64 playerGuid, uint32
     }
 }
 
-bool ChatHandler::HandleCharacterLevelCommand(const char *args)
-{
-    char* nameStr;
-    char* levelStr;
-    extractOptFirstArg((char*)args, &nameStr, &levelStr);
-    if (!levelStr)
-        return false;
-
-    // exception opt second arg: .character level $name
-    if (isalpha(levelStr[0]))
-    {
-        nameStr = levelStr;
-        levelStr = NULL;                                    // current level will used
-    }
-
-    Player* target;
-    uint64 target_guid;
-    std::string target_name;
-    if (!extractPlayerTarget(nameStr, &target, &target_guid, &target_name))
-        return false;
-
-    int32 oldlevel = target ? target->getLevel() : Player::GetLevelFromDB(target_guid);
-    int32 newlevel = levelStr ? atoi(levelStr) : oldlevel;
-
-    if (newlevel < 1)
-        return false;                                       // invalid level
-
-    if (newlevel > STRONG_MAX_LEVEL)                         // hardcoded maximum level
-        newlevel = STRONG_MAX_LEVEL;
-
-    HandleCharacterLevel(target, target_guid, oldlevel, newlevel);
-
-    if (!m_session || m_session->GetPlayer() != target)      // including player == NULL
-    {
-        std::string nameLink = playerLink(target_name);
-        PSendSysMessage(LANG_YOU_CHANGE_LVL, nameLink.c_str(), newlevel);
-    }
-
-    return true;
-}
-
 bool ChatHandler::HandleLevelUpCommand(const char *args)
 {
     char* nameStr;
@@ -2695,152 +2654,6 @@ bool ChatHandler::HandleResetAllCommand(const char * args)
     for (HashMapHolder<Player>::MapType::const_iterator itr = plist.begin(); itr != plist.end(); ++itr)
         itr->second->SetAtLoginFlag(atLogin);
 
-    return true;
-}
-
-bool ChatHandler::HandleServerShutDownCancelCommand(const char* /*args*/)
-{
-    sWorld->ShutdownCancel();
-    return true;
-}
-
-bool ChatHandler::HandleServerShutDownCommand(const char *args)
-{
-    if (!*args)
-        return false;
-
-    char* time_str = strtok ((char*) args, " ");
-    char* exitcode_str = strtok (NULL, "");
-
-    int32 time = atoi (time_str);
-
-    ///- Prevent interpret wrong arg value as 0 secs shutdown time
-    if ((time == 0 && (time_str[0] != '0' || time_str[1] != '\0')) || time < 0)
-        return false;
-
-    if (exitcode_str)
-    {
-        int32 exitcode = atoi (exitcode_str);
-
-        // Handle atoi() errors
-        if (exitcode == 0 && (exitcode_str[0] != '0' || exitcode_str[1] != '\0'))
-            return false;
-
-        // Exit code should be in range of 0-125, 126-255 is used
-        // in many shells for their own return codes and code > 255
-        // is not supported in many others
-        if (exitcode < 0 || exitcode > 125)
-            return false;
-
-        sWorld->ShutdownServ(time, 0, exitcode);
-    }
-    else
-        sWorld->ShutdownServ(time, 0, SHUTDOWN_EXIT_CODE);
-    return true;
-}
-
-bool ChatHandler::HandleServerRestartCommand(const char *args)
-{
-    if (!*args)
-        return false;
-
-    char* time_str = strtok ((char*) args, " ");
-    char* exitcode_str = strtok (NULL, "");
-
-    int32 time = atoi (time_str);
-
-    ///- Prevent interpret wrong arg value as 0 secs shutdown time
-    if ((time == 0 && (time_str[0] != '0' || time_str[1] != '\0')) || time < 0)
-        return false;
-
-    if (exitcode_str)
-    {
-        int32 exitcode = atoi (exitcode_str);
-
-        // Handle atoi() errors
-        if (exitcode == 0 && (exitcode_str[0] != '0' || exitcode_str[1] != '\0'))
-            return false;
-
-        // Exit code should be in range of 0-125, 126-255 is used
-        // in many shells for their own return codes and code > 255
-        // is not supported in many others
-        if (exitcode < 0 || exitcode > 125)
-            return false;
-
-        sWorld->ShutdownServ(time, SHUTDOWN_MASK_RESTART, exitcode);
-    }
-    else
-        sWorld->ShutdownServ(time, SHUTDOWN_MASK_RESTART, RESTART_EXIT_CODE);
-    return true;
-}
-
-bool ChatHandler::HandleServerIdleRestartCommand(const char *args)
-{
-    if (!*args)
-        return false;
-
-    char* time_str = strtok ((char*) args, " ");
-    char* exitcode_str = strtok (NULL, "");
-
-    int32 time = atoi (time_str);
-
-    ///- Prevent interpret wrong arg value as 0 secs shutdown time
-    if ((time == 0 && (time_str[0] != '0' || time_str[1] != '\0')) || time < 0)
-        return false;
-
-    if (exitcode_str)
-    {
-        int32 exitcode = atoi (exitcode_str);
-
-        // Handle atoi() errors
-        if (exitcode == 0 && (exitcode_str[0] != '0' || exitcode_str[1] != '\0'))
-            return false;
-
-        // Exit code should be in range of 0-125, 126-255 is used
-        // in many shells for their own return codes and code > 255
-        // is not supported in many others
-        if (exitcode < 0 || exitcode > 125)
-            return false;
-
-        sWorld->ShutdownServ(time, SHUTDOWN_MASK_RESTART|SHUTDOWN_MASK_IDLE, exitcode);
-    }
-    else
-        sWorld->ShutdownServ(time, SHUTDOWN_MASK_RESTART|SHUTDOWN_MASK_IDLE, RESTART_EXIT_CODE);
-    return true;
-}
-
-bool ChatHandler::HandleServerIdleShutDownCommand(const char *args)
-{
-    if (!*args)
-        return false;
-
-    char* time_str = strtok ((char*) args, " ");
-    char* exitcode_str = strtok (NULL, "");
-
-    int32 time = atoi (time_str);
-
-    ///- Prevent interpret wrong arg value as 0 secs shutdown time
-    if ((time == 0 && (time_str[0] != '0' || time_str[1] != '\0')) || time < 0)
-        return false;
-
-    if (exitcode_str)
-    {
-        int32 exitcode = atoi (exitcode_str);
-
-        // Handle atoi() errors
-        if (exitcode == 0 && (exitcode_str[0] != '0' || exitcode_str[1] != '\0'))
-            return false;
-
-        // Exit code should be in range of 0-125, 126-255 is used
-        // in many shells for their own return codes and code > 255
-        // is not supported in many others
-        if (exitcode < 0 || exitcode > 125)
-            return false;
-
-        sWorld->ShutdownServ(time, SHUTDOWN_MASK_IDLE, exitcode);
-    }
-    else
-        sWorld->ShutdownServ(time, SHUTDOWN_MASK_IDLE, SHUTDOWN_EXIT_CODE);
     return true;
 }
 
@@ -3764,56 +3577,6 @@ bool ChatHandler::HandleMovegensCommand(const char* /*args*/)
     return true;
 }
 
-bool ChatHandler::HandleServerPLimitCommand(const char *args)
-{
-    if (*args)
-    {
-        char* param = strtok((char*)args, " ");
-        if (!param)
-            return false;
-
-        int l = strlen(param);
-
-        if (strncmp(param, "player", l) == 0)
-            sWorld->SetPlayerSecurityLimit(SEC_PLAYER);
-        else if (strncmp(param, "moderator", l) == 0)
-            sWorld->SetPlayerSecurityLimit(SEC_MODERATOR);
-        else if (strncmp(param, "gamemaster", l) == 0)
-            sWorld->SetPlayerSecurityLimit(SEC_GAMEMASTER);
-        else if (strncmp(param, "administrator", l) == 0)
-            sWorld->SetPlayerSecurityLimit(SEC_ADMINISTRATOR);
-        else if (strncmp(param, "reset", l) == 0)
-        {
-            sWorld->SetPlayerAmountLimit(ConfigMgr::GetIntDefault("PlayerLimit", 100));
-            sWorld->LoadDBAllowedSecurityLevel();
-        }
-        else
-        {
-            int val = atoi(param);
-            if (val < 0)
-                sWorld->SetPlayerSecurityLimit(AccountTypes(uint32(-val)));
-            else
-                sWorld->SetPlayerAmountLimit(uint32(val));
-        }
-    }
-
-    uint32 pLimit = sWorld->GetPlayerAmountLimit();
-    AccountTypes allowedAccountType = sWorld->GetPlayerSecurityLimit();
-    char const* secName = "";
-    switch (allowedAccountType)
-    {
-        case SEC_PLAYER:        secName = "Player";        break;
-        case SEC_MODERATOR:     secName = "Moderator";     break;
-        case SEC_GAMEMASTER:    secName = "Gamemaster";    break;
-        case SEC_ADMINISTRATOR: secName = "Administrator"; break;
-        default:                secName = "<unknown>";     break;
-    }
-
-    PSendSysMessage("Player limits: amount %u, min. security level %s.", pLimit, secName);
-
-    return true;
-}
-
 /*
 ComeToMe command REQUIRED for 3rd party scripting library to have access to PointMovementGenerator
 Without this function 3rd party scripting library will get linking errors (unresolved external)
@@ -3838,35 +3601,6 @@ bool ChatHandler::HandleComeToMeCommand(const char *args)
 
     caster->GetMotionMaster()->MovePoint(0, player->GetPositionX(), player->GetPositionY(), player->GetPositionZ());
     return true;
-}
-
-/// Define the 'Message of the day' for the realm
-bool ChatHandler::HandleServerSetMotdCommand(const char *args)
-{
-    sWorld->SetMotd(args);
-    PSendSysMessage(LANG_MOTD_NEW, args);
-    return true;
-}
-
-/// Set whether we accept new clients
-bool ChatHandler::HandleServerSetClosedCommand(const char *args)
-{
-    if (strncmp(args, "on", 3) == 0)
-    {
-        SendSysMessage(LANG_WORLD_CLOSED);
-        sWorld->SetClosed(true);
-        return true;
-    }
-    else if (strncmp(args, "off", 4) == 0)
-    {
-        SendSysMessage(LANG_WORLD_OPENED);
-        sWorld->SetClosed(false);
-        return true;
-    }
-
-    SendSysMessage(LANG_USE_BOL);
-    SetSentErrorMessage(true);
-    return false;
 }
 
 //Send items by mail
