+++ src/server/scripts/Commands/CMakeLists.txt
@@ -27,9 +27,10 @@ set(scripts_STAT_SRCS
   Commands/cs_quest.cpp
   Commands/cs_reload.cpp
   Commands/cs_tele.cpp
+  Commands/cs_server.cpp
   Commands/cs_titles.cpp
   Commands/cs_wp.cpp
-#  Commands/cs_character.cpp
+  Commands/cs_character.cpp
 #  Commands/cs_list.cpp
 #  Commands/cs_lookup.cpp
 #  Commands/cs_pdump.cpp
