+++ src/server/worldserver/CommandLine/CliRunnable.cpp
@@ -116,458 +116,6 @@ void commandFinished(void*, bool /*success*/)
     fflush(stdout);
 }
 
-/**
- * Collects all GUIDs (and related info) from deleted characters which are still in the database.
- *
- * @param foundList    a reference to an std::list which will be filled with info data
- * @param searchString the search string which either contains a player GUID or a part fo the character-name
- * @return             returns false if there was a problem while selecting the characters (e.g. player name not normalizeable)
- */
-bool ChatHandler::GetDeletedCharacterInfoList(DeletedInfoList& foundList, std::string searchString)
-{
-    PreparedQueryResult result;
-    PreparedStatement* stmt;
-    if (!searchString.empty())
-    {
-        // search by GUID
-        if (isNumeric(searchString.c_str()))
-        {
-            stmt = CharacterDatabase.GetPreparedStatement(CHAR_SEL_CHAR_DEL_INFO_BY_GUID);
-
-            stmt->setUInt32(0, uint32(atoi(searchString.c_str())));
-
-            result = CharacterDatabase.Query(stmt);
-        }
-        // search by name
-        else
-        {
-            if (!normalizePlayerName(searchString))
-                return false;
-
-            stmt = CharacterDatabase.GetPreparedStatement(CHAR_SEL_CHAR_DEL_INFO_BY_NAME);
-
-            stmt->setString(0, searchString);
-
-            result = CharacterDatabase.Query(stmt);
-        }
-    }
-    else
-    {
-        stmt = CharacterDatabase.GetPreparedStatement(CHAR_SEL_CHAR_DEL_INFO);
-
-        result = CharacterDatabase.Query(stmt);
-    }
-
-    if (result)
-    {
-        do
-        {
-            Field* fields = result->Fetch();
-
-            DeletedInfo info;
-
-            info.lowguid    = fields[0].GetUInt32();
-            info.name       = fields[1].GetString();
-            info.accountId  = fields[2].GetUInt32();
-
-            // account name will be empty for not existed account
-            AccountMgr::GetName(info.accountId, info.accountName);
-
-            info.deleteDate = time_t(fields[3].GetUInt32());
-
-            foundList.push_back(info);
-        } while (result->NextRow());
-    }
-
-    return true;
-}
-
-/**
- * Generate WHERE guids list by deleted info in way preventing return too long where list for existed query string length limit.
- *
- * @param itr          a reference to an deleted info list iterator, it updated in function for possible next function call if list to long
- * @param itr_end      a reference to an deleted info list iterator end()
- * @return             returns generated where list string in form: 'guid IN (gui1, guid2, ...)'
- */
-std::string ChatHandler::GenerateDeletedCharacterGUIDsWhereStr(DeletedInfoList::const_iterator& itr, DeletedInfoList::const_iterator const& itr_end)
-{
-    std::ostringstream wherestr;
-    wherestr << "guid IN ('";
-    for (; itr != itr_end; ++itr)
-    {
-        wherestr << itr->lowguid;
-
-        if (wherestr.str().size() > MAX_QUERY_LEN - 50)     // near to max query
-        {
-            ++itr;
-            break;
-        }
-
-        DeletedInfoList::const_iterator itr2 = itr;
-        if (++itr2 != itr_end)
-            wherestr << "', '";
-    }
-    wherestr << "')";
-    return wherestr.str();
-}
-
-/**
- * Shows all deleted characters which matches the given search string, expected non empty list
- *
- * @see ChatHandler::HandleCharacterDeletedListCommand
- * @see ChatHandler::HandleCharacterDeletedRestoreCommand
- * @see ChatHandler::HandleCharacterDeletedDeleteCommand
- * @see ChatHandler::DeletedInfoList
- *
- * @param foundList contains a list with all found deleted characters
- */
-void ChatHandler::HandleCharacterDeletedListHelper(DeletedInfoList const& foundList)
-{
-    if (!m_session)
-    {
-        SendSysMessage(LANG_CHARACTER_DELETED_LIST_BAR);
-        SendSysMessage(LANG_CHARACTER_DELETED_LIST_HEADER);
-        SendSysMessage(LANG_CHARACTER_DELETED_LIST_BAR);
-    }
-
-    for (DeletedInfoList::const_iterator itr = foundList.begin(); itr != foundList.end(); ++itr)
-    {
-        std::string dateStr = TimeToTimestampStr(itr->deleteDate);
-
-        if (!m_session)
-            PSendSysMessage(LANG_CHARACTER_DELETED_LIST_LINE_CONSOLE,
-                itr->lowguid, itr->name.c_str(), itr->accountName.empty() ? "<Not existed>" : itr->accountName.c_str(),
-                itr->accountId, dateStr.c_str());
-        else
-            PSendSysMessage(LANG_CHARACTER_DELETED_LIST_LINE_CHAT,
-                itr->lowguid, itr->name.c_str(), itr->accountName.empty() ? "<Not existed>" : itr->accountName.c_str(),
-                itr->accountId, dateStr.c_str());
-    }
-
-    if (!m_session)
-        SendSysMessage(LANG_CHARACTER_DELETED_LIST_BAR);
-}
-
-/**
- * Handles the '.character deleted list' command, which shows all deleted characters which matches the given search string
- *
- * @see ChatHandler::HandleCharacterDeletedListHelper
- * @see ChatHandler::HandleCharacterDeletedRestoreCommand
- * @see ChatHandler::HandleCharacterDeletedDeleteCommand
- * @see ChatHandler::DeletedInfoList
- *
- * @param args the search string which either contains a player GUID or a part fo the character-name
- */
-bool ChatHandler::HandleCharacterDeletedListCommand(const char* args)
-{
-    DeletedInfoList foundList;
-    if (!GetDeletedCharacterInfoList(foundList, args))
-        return false;
-
-    // if no characters have been found, output a warning
-    if (foundList.empty())
-    {
-        SendSysMessage(LANG_CHARACTER_DELETED_LIST_EMPTY);
-        return false;
-    }
-
-    HandleCharacterDeletedListHelper(foundList);
-    return true;
-}
-
-/**
- * Restore a previously deleted character
- *
- * @see ChatHandler::HandleCharacterDeletedListHelper
- * @see ChatHandler::HandleCharacterDeletedRestoreCommand
- * @see ChatHandler::HandleCharacterDeletedDeleteCommand
- * @see ChatHandler::DeletedInfoList
- *
- * @param delInfo the informations about the character which will be restored
- */
-void ChatHandler::HandleCharacterDeletedRestoreHelper(DeletedInfo const& delInfo)
-{
-    if (delInfo.accountName.empty())                    // account not exist
-    {
-        PSendSysMessage(LANG_CHARACTER_DELETED_SKIP_ACCOUNT, delInfo.name.c_str(), delInfo.lowguid, delInfo.accountId);
-        return;
-    }
-
-    // check character count
-    uint32 charcount = AccountMgr::GetCharactersCount(delInfo.accountId);
-    if (charcount >= 10)
-    {
-        PSendSysMessage(LANG_CHARACTER_DELETED_SKIP_FULL, delInfo.name.c_str(), delInfo.lowguid, delInfo.accountId);
-        return;
-    }
-
-    if (sObjectMgr->GetPlayerGUIDByName(delInfo.name))
-    {
-        PSendSysMessage(LANG_CHARACTER_DELETED_SKIP_NAME, delInfo.name.c_str(), delInfo.lowguid, delInfo.accountId);
-        return;
-    }
-
-    PreparedStatement* stmt = CharacterDatabase.GetPreparedStatement(CHAR_UDP_RESTORE_DELETE_INFO);
-
-    stmt->setString(0, delInfo.name);
-    stmt->setUInt32(1, delInfo.accountId);
-    stmt->setUInt32(2, delInfo.lowguid);
-
-    CharacterDatabase.Execute(stmt);
-
-    stmt = CharacterDatabase.GetPreparedStatement(CHAR_SEL_CHARACTER_NAME_DATA);
-    stmt->setUInt32(0, delInfo.lowguid);
-    if (PreparedQueryResult result = CharacterDatabase.Query(stmt))
-        sWorld->AddCharacterNameData(delInfo.lowguid, delInfo.name, (*result)[2].GetUInt8(), (*result)[0].GetUInt8(), (*result)[1].GetUInt8());
-}
-
-/**
- * Handles the '.character deleted restore' command, which restores all deleted characters which matches the given search string
- *
- * The command automatically calls '.character deleted list' command with the search string to show all restored characters.
- *
- * @see ChatHandler::HandleCharacterDeletedRestoreHelper
- * @see ChatHandler::HandleCharacterDeletedListCommand
- * @see ChatHandler::HandleCharacterDeletedDeleteCommand
- *
- * @param args the search string which either contains a player GUID or a part of the character-name
- */
-bool ChatHandler::HandleCharacterDeletedRestoreCommand(const char* args)
-{
-    // It is required to submit at least one argument
-    if (!*args)
-        return false;
-
-    std::string searchString;
-    std::string newCharName;
-    uint32 newAccount = 0;
-
-    // GCC by some strange reason fail build code without temporary variable
-    std::istringstream params(args);
-    params >> searchString >> newCharName >> newAccount;
-
-    DeletedInfoList foundList;
-    if (!GetDeletedCharacterInfoList(foundList, searchString))
-        return false;
-
-    if (foundList.empty())
-    {
-        SendSysMessage(LANG_CHARACTER_DELETED_LIST_EMPTY);
-        return false;
-    }
-
-    SendSysMessage(LANG_CHARACTER_DELETED_RESTORE);
-    HandleCharacterDeletedListHelper(foundList);
-
-    if (newCharName.empty())
-    {
-        // Drop not existed account cases
-        for (DeletedInfoList::iterator itr = foundList.begin(); itr != foundList.end(); ++itr)
-            HandleCharacterDeletedRestoreHelper(*itr);
-    }
-    else if (foundList.size() == 1 && normalizePlayerName(newCharName))
-    {
-        DeletedInfo delInfo = foundList.front();
-
-        // update name
-        delInfo.name = newCharName;
-
-        // if new account provided update deleted info
-        if (newAccount && newAccount != delInfo.accountId)
-        {
-            delInfo.accountId = newAccount;
-            AccountMgr::GetName(newAccount, delInfo.accountName);
-        }
-
-        HandleCharacterDeletedRestoreHelper(delInfo);
-    }
-    else
-        SendSysMessage(LANG_CHARACTER_DELETED_ERR_RENAME);
-
-    return true;
-}
-
-/**
- * Handles the '.character deleted delete' command, which completely deletes all deleted characters which matches the given search string
- *
- * @see Player::GetDeletedCharacterGUIDs
- * @see Player::DeleteFromDB
- * @see ChatHandler::HandleCharacterDeletedListCommand
- * @see ChatHandler::HandleCharacterDeletedRestoreCommand
- *
- * @param args the search string which either contains a player GUID or a part fo the character-name
- */
-bool ChatHandler::HandleCharacterDeletedDeleteCommand(const char* args)
-{
-    // It is required to submit at least one argument
-    if (!*args)
-        return false;
-
-    DeletedInfoList foundList;
-    if (!GetDeletedCharacterInfoList(foundList, args))
-        return false;
-
-    if (foundList.empty())
-    {
-        SendSysMessage(LANG_CHARACTER_DELETED_LIST_EMPTY);
-        return false;
-    }
-
-    SendSysMessage(LANG_CHARACTER_DELETED_DELETE);
-    HandleCharacterDeletedListHelper(foundList);
-
-    // Call the appropriate function to delete them (current account for deleted characters is 0)
-    for (DeletedInfoList::const_iterator itr = foundList.begin(); itr != foundList.end(); ++itr)
-        Player::DeleteFromDB(itr->lowguid, 0, false, true);
-
-    return true;
-}
-
-/**
- * Handles the '.character deleted old' command, which completely deletes all deleted characters deleted with some days ago
- *
- * @see Player::DeleteOldCharacters
- * @see Player::DeleteFromDB
- * @see ChatHandler::HandleCharacterDeletedDeleteCommand
- * @see ChatHandler::HandleCharacterDeletedListCommand
- * @see ChatHandler::HandleCharacterDeletedRestoreCommand
- *
- * @param args the search string which either contains a player GUID or a part fo the character-name
- */
-bool ChatHandler::HandleCharacterDeletedOldCommand(const char* args)
-{
-    int32 keepDays = sWorld->getIntConfig(CONFIG_CHARDELETE_KEEP_DAYS);
-
-    char* px = strtok((char*)args, " ");
-    if (px)
-    {
-        if (!isNumeric(px))
-            return false;
-
-        keepDays = atoi(px);
-        if (keepDays < 0)
-            return false;
-    }
-    // config option value 0 -> disabled and can't be used
-    else if (keepDays <= 0)
-        return false;
-
-    Player::DeleteOldCharacters((uint32)keepDays);
-    return true;
-}
-
-bool ChatHandler::HandleCharacterEraseCommand(const char* args){
-    if (!*args)
-        return false;
-
-    char *character_name_str = strtok((char*)args, " ");
-    if (!character_name_str)
-        return false;
-
-    std::string character_name = character_name_str;
-    if (!normalizePlayerName(character_name))
-        return false;
-
-    uint64 character_guid;
-    uint32 account_id;
-
-    Player* player = sObjectAccessor->FindPlayerByName(character_name.c_str());
-    if (player)
-    {
-        character_guid = player->GetGUID();
-        account_id = player->GetSession()->GetAccountId();
-        player->GetSession()->KickPlayer();
-    }
-    else
-    {
-        character_guid = sObjectMgr->GetPlayerGUIDByName(character_name);
-        if (!character_guid)
-        {
-            PSendSysMessage(LANG_NO_PLAYER, character_name.c_str());
-            SetSentErrorMessage(true);
-            return false;
-        }
-
-        account_id = sObjectMgr->GetPlayerAccountIdByGUID(character_guid);
-    }
-
-    std::string account_name;
-    AccountMgr::GetName (account_id, account_name);
-
-    Player::DeleteFromDB(character_guid, account_id, true, true);
-    PSendSysMessage(LANG_CHARACTER_DELETED, character_name.c_str(), GUID_LOPART(character_guid), account_name.c_str(), account_id);
-    return true;
-}
-
-/// Exit the realm
-bool ChatHandler::HandleServerExitCommand(const char* /*args*/)
-{
-    SendSysMessage(LANG_COMMAND_EXIT);
-    World::StopNow(SHUTDOWN_EXIT_CODE);
-    return true;
-}
-
-/// Set the level of logging
-bool ChatHandler::HandleServerSetLogFileLevelCommand(const char *args)
-{
-    if (!*args)
-        return false;
-
-    char *NewLevel = strtok((char*)args, " ");
-    if (!NewLevel)
-        return false;
-
-    sLog->SetLogFileLevel(NewLevel);
-    return true;
-}
-
-/// Set the level of logging
-bool ChatHandler::HandleServerSetLogLevelCommand(const char *args)
-{
-    if (!*args)
-        return false;
-
-    char *NewLevel = strtok((char*)args, " ");
-    if (!NewLevel)
-        return false;
-
-    sLog->SetLogLevel(NewLevel);
-    return true;
-}
-
-/// set diff time record interval
-bool ChatHandler::HandleServerSetDiffTimeCommand(const char *args)
-{
-    if (!*args)
-        return false;
-
-    char *NewTimeStr = strtok((char*)args, " ");
-    if (!NewTimeStr)
-        return false;
-
-    int32 NewTime =atoi(NewTimeStr);
-    if (NewTime < 0)
-        return false;
-
-    sWorld->SetRecordDiffInterval(NewTime);
-    printf( "Record diff every %u ms\n", NewTime);
-    return true;
-}
-
-/// toggle sql driver query logging
-bool ChatHandler::HandleServerToggleQueryLogging(const char* /* args */)
-{
-    sLog->SetSQLDriverQueryLogging(!sLog->GetSQLDriverQueryLogging());
-    if (sLog->GetSQLDriverQueryLogging())
-        PSendSysMessage(LANG_SQLDRIVER_QUERY_LOGGING_ENABLED);
-    else
-        PSendSysMessage(LANG_SQLDRIVER_QUERY_LOGGING_DISABLED);
-
-    return true;
-}
-
-/// @}
-
 #ifdef linux
 // Non-blocking keypress detector, when return pressed, return 1, else always return 0
 int kb_hit_return()
